editable module
