+++ views/traveler-config.jade
@@ -104,4 +104,4 @@ block js
   script(type='text/javascript', src='#{prefix}/dependencies/moment.js')
   script(type='text/javascript', src='#{prefix}/bootstrap-datepicker/js/bootstrap-datepicker.js')
   script(type='text/javascript', src='#{prefix}/javascripts/ajax-helper.js')
-  script(type='text/javascript', src='#{prefix}/javascripts/traveler-config.js')
+  script(type='module', src='#{prefix}/javascripts/traveler-config.js')
